Assert classpath structure instead of elapsed time in re-export tests
The 5s and 10s wall-clock budgets were meaningless on shared CI hardware.
Both tests now assert on the computed classpath entries and access rules.

Fixes https://github.com/eclipse-pde/eclipse.pde/issues/2434
