+++ ui/org.eclipse.pde.ui.tests/src/org/eclipse/pde/core/tests/internal/classpath/CyclicReexportClasspathTest.java
@@ -13,8 +13,7 @@
  *******************************************************************************/
 package org.eclipse.pde.core.tests.internal.classpath;
 
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -28,6 +27,7 @@ import java.util.jar.JarOutputStream;
 import java.util.jar.Manifest;
 import java.util.zip.ZipEntry;
 
+import org.eclipse.core.resources.IFile;
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.resources.IWorkspaceDescription;
 import org.eclipse.core.resources.IncrementalProjectBuilder;
@@ -35,16 +35,19 @@ import org.eclipse.core.resources.ResourcesPlugin;
 import org.eclipse.core.runtime.NullProgressMonitor;
 import org.eclipse.jdt.core.IClasspathEntry;
 import org.eclipse.jdt.core.JavaCore;
+import org.eclipse.pde.core.build.IBuild;
+import org.eclipse.pde.core.build.IBuildEntry;
 import org.eclipse.pde.core.plugin.IPluginModelBase;
 import org.eclipse.pde.core.plugin.PluginRegistry;
 import org.eclipse.pde.core.project.IBundleProjectDescription;
 import org.eclipse.pde.core.project.IBundleProjectService;
-import org.eclipse.pde.core.project.IRequiredBundleDescription;
 import org.eclipse.pde.core.target.ITargetDefinition;
 import org.eclipse.pde.core.target.ITargetLocation;
 import org.eclipse.pde.core.target.ITargetPlatformService;
+import org.eclipse.pde.internal.core.ClasspathComputer;
 import org.eclipse.pde.internal.core.PDECore;
-import org.eclipse.pde.internal.ui.wizards.tools.UpdateClasspathJob;
+import org.eclipse.pde.internal.core.build.WorkspaceBuildModel;
+import org.eclipse.pde.internal.core.project.PDEProject;
 import org.eclipse.pde.ui.tests.runtime.TestUtils;
 import org.eclipse.pde.ui.tests.util.ProjectUtils;
 import org.eclipse.pde.ui.tests.util.TargetPlatformUtil;
@@ -54,13 +57,11 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.osgi.framework.Version;
-import org.osgi.framework.VersionRange;
 
-public class ChainedReexportPerformanceTest {
+/** Classpath of a re-export cycle in a secondary dependency. */
+public class CyclicReexportClasspathTest {
 
-	private static final String CHAIN_PREFIX = "Chain_";
-	private static final int PACKAGE_COUNT = 1000;
-	private static final int BUNDLE_CHAIN_DEPTH = 5;
+	private static final String CYCLE_BUNDLE_PREFIX = "Cycle_";
 	private File targetDir;
 	private ITargetDefinition savedTarget;
 
@@ -82,8 +83,8 @@ public class ChainedReexportPerformanceTest {
 		desc.setAutoBuilding(false);
 		ResourcesPlugin.getWorkspace().setDescription(desc);
 
-		targetDir = Files.createTempDirectory("pde_chain_perf_target").toFile();
-		createChainedTargetPlatform();
+		targetDir = Files.createTempDirectory("pde_perf_target").toFile();
+		createCyclicTargetPlatform();
 	}
 
 	@AfterEach
@@ -101,30 +102,18 @@ public class ChainedReexportPerformanceTest {
 		}
 	}
 
-	private void createChainedTargetPlatform() throws Exception {
-		// Create a chain of bundles: B_0 -> B_1 -> ... -> B_N (all re-exporting)
-		for (int i = 0; i < BUNDLE_CHAIN_DEPTH; i++) {
-			String name = CHAIN_PREFIX + i;
-			String exports = createPackageExports(name);
-			String requires = (i > 0) ? (CHAIN_PREFIX + (i - 1) + ";visibility:=reexport") : null;
-			createBundle(targetDir, name, exports, requires);
-		}
+	private void createCyclicTargetPlatform() throws Exception {
+		// Cycle_A -> reexports Cycle_B
+		// Cycle_B -> reexports Cycle_C
+		// Cycle_C -> reexports Cycle_A
+		createBundle(targetDir, CYCLE_BUNDLE_PREFIX + "A", null, CYCLE_BUNDLE_PREFIX + "B;visibility:=reexport");
+		createBundle(targetDir, CYCLE_BUNDLE_PREFIX + "B", null, CYCLE_BUNDLE_PREFIX + "C;visibility:=reexport");
+		createBundle(targetDir, CYCLE_BUNDLE_PREFIX + "C", null, CYCLE_BUNDLE_PREFIX + "A;visibility:=reexport");
 
 		ITargetPlatformService tps = PDECore.getDefault().acquireService(ITargetPlatformService.class);
 		ITargetDefinition target = tps.newTarget();
 		target.setTargetLocations(new ITargetLocation[] { tps.newDirectoryLocation(targetDir.getAbsolutePath()) });
 		TargetPlatformUtil.loadAndSetTarget(target);
-	}
-
-	private String createPackageExports(String bundleName) {
-		StringBuilder sb = new StringBuilder();
-		for (int i = 0; i < PACKAGE_COUNT; i++) {
-			if (sb.length() > 0) {
-				sb.append(",");
-			}
-			sb.append(bundleName).append(".pkg.").append(i);
-		}
-		return sb.toString();
 	}
 
 	private void createBundle(File dir, String name, String exports, String requires) throws IOException {
@@ -151,7 +140,7 @@ public class ChainedReexportPerformanceTest {
 	}
 
 	@Test
-	public void testChainedReexportPerformance() throws Exception {
+	public void testCyclicReexportVisitsEachBundleOnce() throws Exception {
 		IBundleProjectService service = PDECore.getDefault().acquireService(IBundleProjectService.class);
 
 		String consumerName = "ConsumerBundle";
@@ -162,14 +151,21 @@ public class ChainedReexportPerformanceTest {
 		IBundleProjectDescription consumerDesc = service.getDescription(consumerProj);
 		consumerDesc.setSymbolicName(consumerName);
 		consumerDesc.setBundleVersion(new Version("1.0.0"));
-
-		// Require the last bundle in the chain with re-export
-		IRequiredBundleDescription mainReq = service.newRequiredBundle(CHAIN_PREFIX + (BUNDLE_CHAIN_DEPTH - 1),
-				(VersionRange) null, false, true);
-		consumerDesc.setRequiredBundles(new IRequiredBundleDescription[] { mainReq });
-
 		consumerDesc.setNatureIds(new String[] { JavaCore.NATURE_ID, IBundleProjectDescription.PLUGIN_NATURE });
 		consumerDesc.apply(null);
+
+		// Add secondary dependency to build.properties
+		IFile buildProps = PDEProject.getBuildProperties(consumerProj);
+		WorkspaceBuildModel buildModel = new WorkspaceBuildModel(buildProps);
+		buildModel.load();
+		IBuild build = buildModel.getBuild();
+		IBuildEntry entry = build.getEntry(IBuildEntry.SECONDARY_DEPENDENCIES);
+		if (entry == null) {
+			entry = buildModel.getFactory().createEntry(IBuildEntry.SECONDARY_DEPENDENCIES);
+			build.add(entry);
+		}
+		entry.addToken(CYCLE_BUNDLE_PREFIX + "A");
+		buildModel.save();
 
 		ResourcesPlugin.getWorkspace().build(IncrementalProjectBuilder.FULL_BUILD, new NullProgressMonitor());
 		TestUtils.waitForJobs("Init", 500, 5000);
@@ -179,11 +175,12 @@ public class ChainedReexportPerformanceTest {
 			throw new IllegalStateException("Consumer model not found");
 		}
 
-		UpdateClasspathJob.scheduleFor(List.of(consumerModel), false);
-		boolean timedOut = TestUtils.waitForJobs("classpath", 100, 10000);
-		assertFalse(timedOut, "Performance regression: classpath computation timed out for chained re-exports");
+		IClasspathEntry[] entries = ClasspathComputer.computeClasspathEntries(consumerModel, consumerProj);
 
-		IClasspathEntry[] resolvedClasspath = JavaCore.create(consumerProj).getRawClasspath();
-		assertTrue(resolvedClasspath.length > 0, "Classpath should not be empty");
+		// Without the visited-check in findExportedPackages the cycle re-enqueues
+		// its bundles forever, so counting the visits is what this asserts.
+		assertEquals(List.of(CYCLE_BUNDLE_PREFIX + "A.jar", CYCLE_BUNDLE_PREFIX + "B.jar", CYCLE_BUNDLE_PREFIX + "C.jar"),
+				ClasspathTestUtils.entryNames(entries, CYCLE_BUNDLE_PREFIX),
+				"Each bundle of the re-export cycle must be visited and added exactly once");
 	}
 }
