+++ ui/org.eclipse.pde.ui.tests/src/org/eclipse/pde/ui/tests/AllPDETests.java
@@ -14,10 +14,10 @@
 package org.eclipse.pde.ui.tests;
 
 import org.eclipse.pde.core.tests.internal.AllPDECoreTests;
-import org.eclipse.pde.core.tests.internal.classpath.ChainedReexportPerformanceTest;
+import org.eclipse.pde.core.tests.internal.classpath.ChainedReexportClasspathTest;
 import org.eclipse.pde.core.tests.internal.classpath.ClasspathResolutionTest;
 import org.eclipse.pde.core.tests.internal.classpath.ClasspathResolutionTest2;
-import org.eclipse.pde.core.tests.internal.classpath.RequiredPluginsClasspathContainerPerformanceTest;
+import org.eclipse.pde.core.tests.internal.classpath.CyclicReexportClasspathTest;
 import org.eclipse.pde.core.tests.internal.core.builders.BundleErrorReporterTest;
 import org.eclipse.pde.core.tests.internal.util.PDESchemaHelperTest;
 import org.eclipse.pde.ui.tests.build.properties.AllValidatorTests;
@@ -62,8 +62,8 @@ import org.junit.platform.suite.api.Suite;
 	BundleRootTests.class, //
 	PluginRegistryTests.class, //
 	ClasspathResolverTest.class, //
-	RequiredPluginsClasspathContainerPerformanceTest.class, //
-	ChainedReexportPerformanceTest.class, //
+	CyclicReexportClasspathTest.class, //
+	ChainedReexportClasspathTest.class, //
 	ClasspathUpdaterTest.class, //
 	PDESchemaHelperTest.class, //
 	ClasspathContributorTest.class, //
